UI: fix circular dependency/unsatisfied reference
- remove appRoutingProviders, due to being called before initialized -> unsatisfied reference
- fixes circulardep of time of use utils

Co-committed-by: Lukas Rieger <73471197+lukasrgr@users.noreply.github.com>
